requirements.txt
markovify
telethon
python-telegram-bot